Updated read me

+++ src/Presentation/TextItem.java
@@ -67,9 +67,9 @@ public class TextItem extends SlideItem {
 				xsize = (int) bounds.getWidth();
 			}
 			if (bounds.getHeight() > 0) {
-				ysize += bounds.getHeight();
+				ysize += (int) bounds.getHeight();
 			}
-			ysize += layout.getLeading() + layout.getDescent();
+			ysize += (int) (layout.getLeading() + layout.getDescent());
 		}
 		return new Rectangle((int) (myStyle.indent*scale), 0, xsize, ysize );
 	}
